Add formatters concept
